Remove kafka 0.8 binder Rearrange code that was moved over to a new module common
