--- spring-cloud-starter-stream-kafka-0.8/src/main/resources/META-INF/spring.provides
@@ -1,1 +1,0 @@
-provides: spring-cloud-starter-stream-kafka-0.8
