--- spring-cloud-stream-binder-kafka-0.8/src/main/resources/META-INF/spring.binders
@@ -1,2 +1,0 @@
-kafka:\
-org.springframework.cloud.stream.binder.kafka.config.KafkaBinderConfiguration
