--- spring-cloud-stream-binder-kafka-0.8/src/test/java/org/springframework/cloud/stream/binder/kafka/KafkaBinderTests.java
@@ -1,819 +1,0 @@
-/*
- * Copyright 2014-2016 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.cloud.stream.binder.kafka;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.Assert.fail;
-
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.LinkedHashMap;
-import java.util.Properties;
-import java.util.UUID;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-
-import org.junit.Before;
-import org.junit.ClassRule;
-import org.junit.Test;
-
-import org.springframework.beans.DirectFieldAccessor;
-import org.springframework.cloud.stream.binder.BinderException;
-import org.springframework.cloud.stream.binder.Binding;
-import org.springframework.cloud.stream.binder.ExtendedConsumerProperties;
-import org.springframework.cloud.stream.binder.ExtendedProducerProperties;
-import org.springframework.cloud.stream.binder.PartitionCapableBinderTests;
-import org.springframework.cloud.stream.binder.Spy;
-import org.springframework.cloud.stream.binder.TestUtils;
-import org.springframework.cloud.stream.binder.kafka.config.KafkaBinderConfigurationProperties;
-import org.springframework.cloud.stream.binder.kafka.config.KafkaConsumerProperties;
-import org.springframework.cloud.stream.binder.kafka.config.KafkaProducerProperties;
-import org.springframework.cloud.stream.binder.test.junit.kafka.KafkaTestSupport;
-import org.springframework.context.support.GenericApplicationContext;
-import org.springframework.integration.channel.DirectChannel;
-import org.springframework.integration.channel.QueueChannel;
-import org.springframework.integration.kafka.core.Partition;
-import org.springframework.integration.kafka.core.TopicNotFoundException;
-import org.springframework.integration.kafka.support.KafkaHeaders;
-import org.springframework.integration.kafka.support.ProducerConfiguration;
-import org.springframework.integration.kafka.support.ProducerMetadata;
-import org.springframework.messaging.Message;
-import org.springframework.messaging.MessageChannel;
-import org.springframework.messaging.MessageHandler;
-import org.springframework.messaging.MessagingException;
-import org.springframework.messaging.support.GenericMessage;
-import org.springframework.messaging.support.MessageBuilder;
-import org.springframework.retry.backoff.FixedBackOffPolicy;
-import org.springframework.retry.policy.SimpleRetryPolicy;
-import org.springframework.retry.support.RetryTemplate;
-
-import kafka.admin.AdminUtils;
-import kafka.api.TopicMetadata;
-
-/**
- * Integration tests for the {@link KafkaMessageChannelBinder}.
- * @author Eric Bottard
- * @author Marius Bogoevici
- * @author Mark Fisher
- * @author Ilayaperumal Gopinathan
- */
-public class KafkaBinderTests extends
-		PartitionCapableBinderTests<KafkaTestBinder, ExtendedConsumerProperties<KafkaConsumerProperties>, ExtendedProducerProperties<KafkaProducerProperties>> {
-
-	private final String CLASS_UNDER_TEST_NAME = KafkaMessageChannelBinder.class.getSimpleName();
-
-	@ClassRule
-	public static KafkaTestSupport kafkaTestSupport = new KafkaTestSupport();
-
-	private KafkaTestBinder binder;
-
-	@Override
-	protected void binderBindUnbindLatency() throws InterruptedException {
-		Thread.sleep(500);
-	}
-
-	@Override
-	protected KafkaTestBinder getBinder() {
-		if (binder == null) {
-			KafkaBinderConfigurationProperties binderConfiguration = createConfigurationProperties();
-			binder = new KafkaTestBinder(binderConfiguration);
-		}
-		return binder;
-	}
-
-	private KafkaBinderConfigurationProperties createConfigurationProperties() {
-		KafkaBinderConfigurationProperties binderConfiguration = new KafkaBinderConfigurationProperties();
-		binderConfiguration.setBrokers(kafkaTestSupport.getBrokerAddress());
-		binderConfiguration.setZkNodes(kafkaTestSupport.getZkConnectString());
-		return binderConfiguration;
-	}
-
-	@Override
-	protected ExtendedConsumerProperties<KafkaConsumerProperties> createConsumerProperties() {
-		return new ExtendedConsumerProperties<>(new KafkaConsumerProperties());
-	}
-
-	@Override
-	protected ExtendedProducerProperties<KafkaProducerProperties> createProducerProperties() {
-		return new ExtendedProducerProperties<>(new KafkaProducerProperties());
-	}
-
-	@Before
-	public void init() {
-		String multiplier = System.getenv("KAFKA_TIMEOUT_MULTIPLIER");
-		if (multiplier != null) {
-			timeoutMultiplier = Double.parseDouble(multiplier);
-		}
-	}
-
-	@Override
-	protected boolean usesExplicitRouting() {
-		return false;
-	}
-
-	@Override
-	protected String getClassUnderTestName() {
-		return CLASS_UNDER_TEST_NAME;
-	}
-
-	@Override
-	public Spy spyOn(final String name) {
-		throw new UnsupportedOperationException("'spyOn' is not used by Kafka tests");
-	}
-
-	@Test
-	public void testDlqAndRetry() throws Exception {
-		KafkaTestBinder binder = getBinder();
-		ExtendedProducerProperties<KafkaProducerProperties> producerProperties = createProducerProperties();
-		producerProperties.setPartitionCount(10);
-		DirectChannel moduleOutputChannel = createBindableChannel("output",
-				createProducerBindingProperties(producerProperties));
-		QueueChannel dlqChannel = new QueueChannel();
-		FailingInvocationCountingMessageHandler handler = new FailingInvocationCountingMessageHandler();
-		ExtendedConsumerProperties<KafkaConsumerProperties> consumerProperties = createConsumerProperties();
-		consumerProperties.setMaxAttempts(3);
-		consumerProperties.setBackOffInitialInterval(100);
-		consumerProperties.setBackOffMaxInterval(150);
-		consumerProperties.getExtension().setEnableDlq(true);
-		DirectChannel moduleInputChannel = createBindableChannel("input", createConsumerBindingProperties(consumerProperties));
-		moduleInputChannel.subscribe(handler);
-		long uniqueBindingId = System.currentTimeMillis();
-		Binding<MessageChannel> producerBinding = binder.bindProducer("retryTest." + uniqueBindingId + ".0",
-				moduleOutputChannel, producerProperties);
-		Binding<MessageChannel> consumerBinding = binder.bindConsumer("retryTest." + uniqueBindingId + ".0",
-				"testGroup", moduleInputChannel, consumerProperties);
-
-		ExtendedConsumerProperties<KafkaConsumerProperties> dlqConsumerProperties = createConsumerProperties();
-		dlqConsumerProperties.setMaxAttempts(1);
-
-		Binding<MessageChannel> dlqConsumerBinding = binder.bindConsumer(
-				"error.retryTest." + uniqueBindingId + ".0.testGroup", null, dlqChannel, dlqConsumerProperties);
-
-		String testMessagePayload = "test." + UUID.randomUUID().toString();
-		Message<String> testMessage = MessageBuilder.withPayload(testMessagePayload).build();
-		moduleOutputChannel.send(testMessage);
-
-		Message<?> receivedMessage = receive(dlqChannel, 3);
-		assertThat(receivedMessage).isNotNull();
-		assertThat(receivedMessage.getPayload()).isEqualTo(testMessagePayload);
-		assertThat(handler.getInvocationCount()).isEqualTo(consumerProperties.getMaxAttempts());
-		dlqConsumerBinding.unbind();
-		consumerBinding.unbind();
-		producerBinding.unbind();
-	}
-
-	@Test
-	public void testDefaultAutoCommitOnErrorWithoutDlq() throws Exception {
-		KafkaTestBinder binder = getBinder();
-		ExtendedProducerProperties<KafkaProducerProperties> producerProperties = createProducerProperties();
-		producerProperties.setPartitionCount(10);
-		DirectChannel moduleOutputChannel = createBindableChannel("output", createProducerBindingProperties(producerProperties));
-		FailingInvocationCountingMessageHandler handler = new FailingInvocationCountingMessageHandler();
-		ExtendedConsumerProperties<KafkaConsumerProperties> consumerProperties = createConsumerProperties();
-		consumerProperties.setMaxAttempts(1);
-		consumerProperties.setBackOffInitialInterval(100);
-		consumerProperties.setBackOffMaxInterval(150);
-		DirectChannel moduleInputChannel = createBindableChannel("input", createConsumerBindingProperties(consumerProperties));
-		moduleInputChannel.subscribe(handler);
-		long uniqueBindingId = System.currentTimeMillis();
-		Binding<MessageChannel> producerBinding = binder.bindProducer("retryTest." + uniqueBindingId + ".0",
-				moduleOutputChannel, producerProperties);
-		Binding<MessageChannel> consumerBinding = binder.bindConsumer("retryTest." + uniqueBindingId + ".0",
-				"testGroup", moduleInputChannel, consumerProperties);
-
-		String testMessagePayload = "test." + UUID.randomUUID().toString();
-		Message<String> testMessage = MessageBuilder.withPayload(testMessagePayload).build();
-		moduleOutputChannel.send(testMessage);
-
-		assertThat(handler.getLatch().await((int) (timeoutMultiplier * 1000), TimeUnit.MILLISECONDS));
-		// first attempt fails
-		assertThat(handler.getReceivedMessages().entrySet()).hasSize(1);
-		Message<?> receivedMessage = handler.getReceivedMessages().entrySet().iterator().next().getValue();
-		assertThat(receivedMessage).isNotNull();
-		assertThat(receivedMessage.getPayload()).isEqualTo(testMessagePayload);
-		assertThat(handler.getInvocationCount()).isEqualTo(consumerProperties.getMaxAttempts());
-		consumerBinding.unbind();
-
-		// on the second attempt the message is redelivered
-		QueueChannel successfulInputChannel = new QueueChannel();
-		consumerBinding = binder.bindConsumer("retryTest." + uniqueBindingId + ".0", "testGroup",
-				successfulInputChannel, consumerProperties);
-		String testMessage2Payload = "test." + UUID.randomUUID().toString();
-		Message<String> testMessage2 = MessageBuilder.withPayload(testMessage2Payload).build();
-		moduleOutputChannel.send(testMessage2);
-
-		Message<?> firstReceived = receive(successfulInputChannel);
-		assertThat(firstReceived.getPayload()).isEqualTo(testMessagePayload);
-		Message<?> secondReceived = receive(successfulInputChannel);
-		assertThat(secondReceived.getPayload()).isEqualTo(testMessage2Payload);
-		consumerBinding.unbind();
-		producerBinding.unbind();
-	}
-
-	@Test
-	public void testDefaultAutoCommitOnErrorWithDlq() throws Exception {
-		KafkaTestBinder binder = getBinder();
-		FailingInvocationCountingMessageHandler handler = new FailingInvocationCountingMessageHandler();
-		ExtendedProducerProperties<KafkaProducerProperties> producerProperties = createProducerProperties();
-		producerProperties.setPartitionCount(10);
-		DirectChannel moduleOutputChannel = createBindableChannel("output", createProducerBindingProperties(producerProperties));
-		ExtendedConsumerProperties<KafkaConsumerProperties> consumerProperties = createConsumerProperties();
-		consumerProperties.setMaxAttempts(3);
-		consumerProperties.setBackOffInitialInterval(100);
-		consumerProperties.setBackOffMaxInterval(150);
-		consumerProperties.getExtension().setEnableDlq(true);
-		DirectChannel moduleInputChannel = createBindableChannel("input", createConsumerBindingProperties(consumerProperties));
-		moduleInputChannel.subscribe(handler);
-		long uniqueBindingId = System.currentTimeMillis();
-		Binding<MessageChannel> producerBinding = binder.bindProducer("retryTest." + uniqueBindingId + ".0",
-				moduleOutputChannel, producerProperties);
-		Binding<MessageChannel> consumerBinding = binder.bindConsumer("retryTest." + uniqueBindingId + ".0",
-				"testGroup", moduleInputChannel, consumerProperties);
-		ExtendedConsumerProperties<KafkaConsumerProperties> dlqConsumerProperties = createConsumerProperties();
-		dlqConsumerProperties.setMaxAttempts(1);
-		QueueChannel dlqChannel = new QueueChannel();
-		Binding<MessageChannel> dlqConsumerBinding = binder.bindConsumer(
-				"error.retryTest." + uniqueBindingId + ".0.testGroup", null, dlqChannel, dlqConsumerProperties);
-
-		String testMessagePayload = "test." + UUID.randomUUID().toString();
-		Message<String> testMessage = MessageBuilder.withPayload(testMessagePayload).build();
-		moduleOutputChannel.send(testMessage);
-
-		Message<?> dlqMessage = receive(dlqChannel, 3);
-		assertThat(dlqMessage).isNotNull();
-		assertThat(dlqMessage.getPayload()).isEqualTo(testMessagePayload);
-
-		// first attempt fails
-		assertThat(handler.getReceivedMessages().entrySet()).hasSize(1);
-		Message<?> handledMessage = handler.getReceivedMessages().entrySet().iterator().next().getValue();
-		assertThat(handledMessage).isNotNull();
-		assertThat(handledMessage.getPayload()).isEqualTo(testMessagePayload);
-		assertThat(handler.getInvocationCount()).isEqualTo(consumerProperties.getMaxAttempts());
-
-		dlqConsumerBinding.unbind();
-		consumerBinding.unbind();
-
-		// on the second attempt the message is not redelivered because the DLQ is set
-		QueueChannel successfulInputChannel = new QueueChannel();
-		consumerBinding = binder.bindConsumer("retryTest." + uniqueBindingId + ".0", "testGroup",
-				successfulInputChannel, consumerProperties);
-		String testMessage2Payload = "test." + UUID.randomUUID().toString();
-		Message<String> testMessage2 = MessageBuilder.withPayload(testMessage2Payload).build();
-		moduleOutputChannel.send(testMessage2);
-
-		Message<?> receivedMessage = receive(successfulInputChannel);
-		assertThat(receivedMessage.getPayload()).isEqualTo(testMessage2Payload);
-
-		consumerBinding.unbind();
-		producerBinding.unbind();
-	}
-
-	@Test(expected = IllegalArgumentException.class)
-	public void testValidateKafkaTopicName() {
-		KafkaTopicUtils.validateTopicName("foo:bar");
-	}
-
-	@Test
-	public void testCompression() throws Exception {
-		final ProducerMetadata.CompressionType[] codecs = new ProducerMetadata.CompressionType[] {
-				ProducerMetadata.CompressionType.none, ProducerMetadata.CompressionType.gzip,
-				ProducerMetadata.CompressionType.snappy };
-		byte[] testPayload = new byte[2048];
-		Arrays.fill(testPayload, (byte) 65);
-		KafkaTestBinder binder = getBinder();
-		for (ProducerMetadata.CompressionType codec : codecs) {
-			ExtendedProducerProperties<KafkaProducerProperties> producerProperties = createProducerProperties();
-			producerProperties.getExtension().setCompressionType(fromProducerMetadataCompressionType(codec));
-
-			DirectChannel moduleOutputChannel = createBindableChannel("input", createProducerBindingProperties(producerProperties));
-			QueueChannel moduleInputChannel = new QueueChannel();
-			Binding<MessageChannel> producerBinding = binder.bindProducer("foo.0", moduleOutputChannel,
-					producerProperties);
-			Binding<MessageChannel> consumerBinding = binder.bindConsumer("foo.0", "test", moduleInputChannel,
-					createConsumerProperties());
-			Message<?> message = org.springframework.integration.support.MessageBuilder.withPayload(testPayload)
-					.build();
-			// Let the consumer actually bind to the producer before sending a msg
-			binderBindUnbindLatency();
-			moduleOutputChannel.send(message);
-			Message<?> inbound = receive(moduleInputChannel);
-			assertThat(inbound).isNotNull();
-			assertThat((byte[]) inbound.getPayload()).containsExactly(testPayload);
-			producerBinding.unbind();
-			consumerBinding.unbind();
-		}
-	}
-
-	@Test
-	public void testCustomPartitionCountOverridesDefaultIfLarger() throws Exception {
-
-		byte[] testPayload = new byte[2048];
-		Arrays.fill(testPayload, (byte) 65);
-		KafkaBinderConfigurationProperties binderConfiguration = createConfigurationProperties();
-		binderConfiguration.setMinPartitionCount(10);
-		KafkaTestBinder binder = new KafkaTestBinder(binderConfiguration);
-
-		ExtendedProducerProperties<KafkaProducerProperties> producerProperties = createProducerProperties();
-		producerProperties.setPartitionCount(10);
-		DirectChannel moduleOutputChannel = createBindableChannel("output", createProducerBindingProperties(producerProperties));
-		QueueChannel moduleInputChannel = new QueueChannel();
-		ExtendedConsumerProperties<KafkaConsumerProperties> consumerProperties = createConsumerProperties();
-		long uniqueBindingId = System.currentTimeMillis();
-		Binding<MessageChannel> producerBinding = binder.bindProducer("foo" + uniqueBindingId + ".0",
-				moduleOutputChannel, producerProperties);
-		Binding<MessageChannel> consumerBinding = binder.bindConsumer("foo" + uniqueBindingId + ".0", null,
-				moduleInputChannel, consumerProperties);
-		Message<?> message = org.springframework.integration.support.MessageBuilder.withPayload(testPayload)
-				.build();
-		// Let the consumer actually bind to the producer before sending a msg
-		binderBindUnbindLatency();
-		moduleOutputChannel.send(message);
-		Message<?> inbound = receive(moduleInputChannel);
-		assertThat(inbound).isNotNull();
-		assertThat((byte[]) inbound.getPayload()).containsExactly(testPayload);
-		Collection<Partition> partitions = binder.getCoreBinder().getConnectionFactory()
-				.getPartitions("foo" + uniqueBindingId + ".0");
-		assertThat(partitions).hasSize(10);
-		producerBinding.unbind();
-		consumerBinding.unbind();
-	}
-
-	@Test
-	public void testCustomPartitionCountDoesNotOverridePartitioningIfSmaller() throws Exception {
-
-		byte[] testPayload = new byte[2048];
-		Arrays.fill(testPayload, (byte) 65);
-		KafkaBinderConfigurationProperties binderConfiguration = createConfigurationProperties();
-		binderConfiguration.setMinPartitionCount(6);
-		KafkaTestBinder binder = new KafkaTestBinder(binderConfiguration);
-		DirectChannel moduleOutputChannel = new DirectChannel();
-		QueueChannel moduleInputChannel = new QueueChannel();
-		ExtendedProducerProperties<KafkaProducerProperties> producerProperties = createProducerProperties();
-		producerProperties.setPartitionCount(5);
-		producerProperties.setPartitionKeyExpression(spelExpressionParser.parseExpression("payload"));
-		ExtendedConsumerProperties<KafkaConsumerProperties> consumerProperties = createConsumerProperties();
-		long uniqueBindingId = System.currentTimeMillis();
-		Binding<MessageChannel> producerBinding = binder.bindProducer("foo" + uniqueBindingId + ".0",
-				moduleOutputChannel, producerProperties);
-		Binding<MessageChannel> consumerBinding = binder.bindConsumer("foo" + uniqueBindingId + ".0", null,
-				moduleInputChannel, consumerProperties);
-		Message<?> message = org.springframework.integration.support.MessageBuilder.withPayload(testPayload)
-				.build();
-		// Let the consumer actually bind to the producer before sending a msg
-		binderBindUnbindLatency();
-		moduleOutputChannel.send(message);
-		Message<?> inbound = receive(moduleInputChannel);
-		assertThat(inbound).isNotNull();
-		assertThat((byte[]) inbound.getPayload()).containsExactly(testPayload);
-		Collection<Partition> partitions = binder.getCoreBinder().getConnectionFactory()
-				.getPartitions("foo" + uniqueBindingId + ".0");
-		assertThat(partitions).hasSize(6);
-		producerBinding.unbind();
-		consumerBinding.unbind();
-	}
-
-	@Test
-	public void testCustomPartitionCountOverridesPartitioningIfLarger() throws Exception {
-
-		byte[] testPayload = new byte[2048];
-		Arrays.fill(testPayload, (byte) 65);
-		KafkaBinderConfigurationProperties binderConfiguration = createConfigurationProperties();
-		binderConfiguration.setMinPartitionCount(4);
-		KafkaTestBinder binder = new KafkaTestBinder(binderConfiguration);
-
-		DirectChannel moduleOutputChannel = new DirectChannel();
-		QueueChannel moduleInputChannel = new QueueChannel();
-		ExtendedProducerProperties<KafkaProducerProperties> producerProperties = createProducerProperties();
-		producerProperties.setPartitionCount(5);
-		producerProperties.setPartitionKeyExpression(spelExpressionParser.parseExpression("payload"));
-		ExtendedConsumerProperties<KafkaConsumerProperties> consumerProperties = createConsumerProperties();
-		long uniqueBindingId = System.currentTimeMillis();
-		Binding<MessageChannel> producerBinding = binder.bindProducer("foo" + uniqueBindingId + ".0",
-				moduleOutputChannel, producerProperties);
-		Binding<MessageChannel> consumerBinding = binder.bindConsumer("foo" + uniqueBindingId + ".0", null,
-				moduleInputChannel, consumerProperties);
-		Message<?> message = org.springframework.integration.support.MessageBuilder.withPayload(testPayload)
-				.build();
-		// Let the consumer actually bind to the producer before sending a msg
-		binderBindUnbindLatency();
-		moduleOutputChannel.send(message);
-		Message<?> inbound = receive(moduleInputChannel);
-		assertThat(inbound).isNotNull();
-		assertThat((byte[]) inbound.getPayload()).containsExactly(testPayload);
-		Collection<Partition> partitions = binder.getCoreBinder().getConnectionFactory()
-				.getPartitions("foo" + uniqueBindingId + ".0");
-		assertThat(partitions).hasSize(5);
-		producerBinding.unbind();
-		consumerBinding.unbind();
-	}
-
-	@Test
-	@SuppressWarnings("unchecked")
-	public void testDefaultConsumerStartsAtEarliest() throws Exception {
-		KafkaMessageChannelBinder binder = new KafkaMessageChannelBinder(createConfigurationProperties());
-		GenericApplicationContext context = new GenericApplicationContext();
-		context.refresh();
-		binder.setApplicationContext(context);
-		binder.afterPropertiesSet();
-		DirectChannel output = new DirectChannel();
-		QueueChannel input1 = new QueueChannel();
-
-		String testTopicName = UUID.randomUUID().toString();
-		binder.bindProducer(testTopicName, output, createProducerProperties());
-		String testPayload1 = "foo-" + UUID.randomUUID().toString();
-		output.send(new GenericMessage<>(testPayload1.getBytes()));
-		binder.bindConsumer(testTopicName, "startOffsets", input1, createConsumerProperties());
-		Message<byte[]> receivedMessage1 = (Message<byte[]>) receive(input1);
-		assertThat(receivedMessage1).isNotNull();
-		assertThat(new String(receivedMessage1.getPayload())).isEqualTo(testPayload1);
-		String testPayload2 = "foo-" + UUID.randomUUID().toString();
-		output.send(new GenericMessage<>(testPayload2.getBytes()));
-		Message<byte[]> receivedMessage2 = (Message<byte[]>) receive(input1);
-		assertThat(receivedMessage2).isNotNull();
-		assertThat(new String(receivedMessage2.getPayload())).isEqualTo(testPayload2);
-	}
-
-	@Test
-	@SuppressWarnings("unchecked")
-	public void testEarliest() throws Exception {
-		KafkaTestBinder binder = getBinder();
-		DirectChannel output = new DirectChannel();
-		QueueChannel input1 = new QueueChannel();
-
-		String testTopicName = UUID.randomUUID().toString();
-		binder.bindProducer(testTopicName, output, createProducerProperties());
-		String testPayload1 = "foo-" + UUID.randomUUID().toString();
-		output.send(new GenericMessage<>(testPayload1.getBytes()));
-		ExtendedConsumerProperties<KafkaConsumerProperties> properties = createConsumerProperties();
-		properties.getExtension().setStartOffset(KafkaConsumerProperties.StartOffset.earliest);
-		binder.bindConsumer(testTopicName, "startOffsets", input1, properties);
-		Message<byte[]> receivedMessage1 = (Message<byte[]>) receive(input1);
-		assertThat(receivedMessage1).isNotNull();
-		String testPayload2 = "foo-" + UUID.randomUUID().toString();
-		output.send(new GenericMessage<>(testPayload2.getBytes()));
-		Message<byte[]> receivedMessage2 = (Message<byte[]>) receive(input1);
-		assertThat(receivedMessage2).isNotNull();
-		assertThat(new String(receivedMessage2.getPayload())).isEqualTo(testPayload2);
-	}
-
-	@Test
-	@SuppressWarnings("unchecked")
-	public void testReset() throws Exception {
-		KafkaTestBinder binder = getBinder();
-		DirectChannel output = new DirectChannel();
-		QueueChannel input1 = new QueueChannel();
-
-		String testTopicName = UUID.randomUUID().toString();
-
-		Binding<MessageChannel> producerBinding = binder.bindProducer(testTopicName, output,
-				createProducerProperties());
-		String testPayload1 = "foo-" + UUID.randomUUID().toString();
-		output.send(new GenericMessage<>(testPayload1.getBytes()));
-		ExtendedConsumerProperties<KafkaConsumerProperties> properties = createConsumerProperties();
-		properties.getExtension().setResetOffsets(true);
-		properties.getExtension().setStartOffset(KafkaConsumerProperties.StartOffset.earliest);
-		Binding<MessageChannel> consumerBinding = binder.bindConsumer(testTopicName, "startOffsets", input1,
-				properties);
-		Message<byte[]> receivedMessage1 = (Message<byte[]>) receive(input1);
-		assertThat(receivedMessage1).isNotNull();
-		String testPayload2 = "foo-" + UUID.randomUUID().toString();
-		output.send(new GenericMessage<>(testPayload2.getBytes()));
-		Message<byte[]> receivedMessage2 = (Message<byte[]>) receive(input1);
-		assertThat(receivedMessage2).isNotNull();
-		assertThat(new String(receivedMessage2.getPayload())).isEqualTo(testPayload2);
-		consumerBinding.unbind();
-
-		String testPayload3 = "foo-" + UUID.randomUUID().toString();
-		output.send(new GenericMessage<>(testPayload3.getBytes()));
-
-		ExtendedConsumerProperties<KafkaConsumerProperties> properties2 = createConsumerProperties();
-		properties2.getExtension().setResetOffsets(true);
-		properties2.getExtension().setStartOffset(KafkaConsumerProperties.StartOffset.earliest);
-		consumerBinding = binder.bindConsumer(testTopicName, "startOffsets", input1, properties2);
-		Message<byte[]> receivedMessage4 = (Message<byte[]>) receive(input1);
-		assertThat(receivedMessage4).isNotNull();
-		assertThat(new String(receivedMessage4.getPayload())).isEqualTo(testPayload1);
-		Message<byte[]> receivedMessage5 = (Message<byte[]>) receive(input1);
-		assertThat(receivedMessage5).isNotNull();
-		assertThat(new String(receivedMessage5.getPayload())).isEqualTo(testPayload2);
-		Message<byte[]> receivedMessage6 = (Message<byte[]>) receive(input1);
-		assertThat(receivedMessage6).isNotNull();
-		assertThat(new String(receivedMessage6.getPayload())).isEqualTo(testPayload3);
-		consumerBinding.unbind();
-		producerBinding.unbind();
-	}
-
-	@Test
-	@SuppressWarnings("unchecked")
-	public void testResume() throws Exception {
-		KafkaBinderConfigurationProperties configurationProperties = createConfigurationProperties();
-		KafkaMessageChannelBinder binder = new KafkaMessageChannelBinder(configurationProperties);
-		GenericApplicationContext context = new GenericApplicationContext();
-		context.refresh();
-		binder.setApplicationContext(context);
-		binder.afterPropertiesSet();
-		DirectChannel output = new DirectChannel();
-		QueueChannel input1 = new QueueChannel();
-
-		String testTopicName = UUID.randomUUID().toString();
-		Binding<MessageChannel> producerBinding = binder.bindProducer(testTopicName, output,
-				createProducerProperties());
-		String testPayload1 = "foo-" + UUID.randomUUID().toString();
-		output.send(new GenericMessage<>(testPayload1.getBytes()));
-		ExtendedConsumerProperties<KafkaConsumerProperties> firstConsumerProperties = createConsumerProperties();
-		Binding<MessageChannel> consumerBinding = binder.bindConsumer(testTopicName, "startOffsets", input1,
-				firstConsumerProperties);
-		Message<byte[]> receivedMessage1 = (Message<byte[]>) receive(input1);
-		assertThat(receivedMessage1).isNotNull();
-		String testPayload2 = "foo-" + UUID.randomUUID().toString();
-		output.send(new GenericMessage<>(testPayload2.getBytes()));
-		Message<byte[]> receivedMessage2 = (Message<byte[]>) receive(input1);
-		assertThat(receivedMessage2).isNotNull();
-		assertThat(new String(receivedMessage2.getPayload())).isNotNull();
-		consumerBinding.unbind();
-
-		String testPayload3 = "foo-" + UUID.randomUUID().toString();
-		output.send(new GenericMessage<>(testPayload3.getBytes()));
-
-		consumerBinding = binder.bindConsumer(testTopicName, "startOffsets", input1, createConsumerProperties());
-		Message<byte[]> receivedMessage3 = (Message<byte[]>) receive(input1);
-		assertThat(receivedMessage3).isNotNull();
-		assertThat(new String(receivedMessage3.getPayload())).isEqualTo(testPayload3);
-		consumerBinding.unbind();
-		producerBinding.unbind();
-	}
-
-	@Test
-	public void testSyncProducerMetadata() throws Exception {
-		KafkaMessageChannelBinder binder = new KafkaMessageChannelBinder(createConfigurationProperties());
-		GenericApplicationContext context = new GenericApplicationContext();
-		context.refresh();
-		binder.setApplicationContext(context);
-		binder.afterPropertiesSet();
-		DirectChannel output = new DirectChannel();
-		String testTopicName = UUID.randomUUID().toString();
-		ExtendedProducerProperties<KafkaProducerProperties> properties = createProducerProperties();
-		properties.getExtension().setSync(true);
-		Binding<MessageChannel> producerBinding = binder.bindProducer(testTopicName, output, properties);
-		DirectFieldAccessor accessor = new DirectFieldAccessor(extractEndpoint(producerBinding));
-		ProducerConfiguration producerConfiguration = (ProducerConfiguration) accessor
-				.getPropertyValue("delegate");
-		assertThat(producerConfiguration.getProducerMetadata().isSync())
-				.withFailMessage("Kafka Sync Producer should have been enabled.");
-		producerBinding.unbind();
-	}
-
-	@Test
-	public void testAutoCreateTopicsDisabledFailsIfTopicMissing() throws Exception {
-		KafkaBinderConfigurationProperties configurationProperties = createConfigurationProperties();
-		configurationProperties.setAutoCreateTopics(false);
-		KafkaMessageChannelBinder binder = new KafkaMessageChannelBinder(configurationProperties);
-		GenericApplicationContext context = new GenericApplicationContext();
-		context.refresh();
-		binder.setApplicationContext(context);
-		binder.afterPropertiesSet();
-		RetryTemplate metatadataRetrievalRetryOperations = new RetryTemplate();
-		metatadataRetrievalRetryOperations.setRetryPolicy(new SimpleRetryPolicy());
-		FixedBackOffPolicy backOffPolicy = new FixedBackOffPolicy();
-		backOffPolicy.setBackOffPeriod(1000);
-		metatadataRetrievalRetryOperations.setBackOffPolicy(backOffPolicy);
-		binder.setMetadataRetryOperations(metatadataRetrievalRetryOperations);
-		ExtendedConsumerProperties<KafkaConsumerProperties> consumerProperties = createConsumerProperties();
-		DirectChannel output = createBindableChannel("output", createConsumerBindingProperties(consumerProperties));
-		String testTopicName = "nonexisting" + System.currentTimeMillis();
-		try {
-			binder.doBindConsumer(testTopicName, "test", output, consumerProperties);
-			fail();
-		}
-		catch (Exception e) {
-			assertThat(e).isInstanceOf(BinderException.class);
-			assertThat(e).hasMessageContaining("Topic " + testTopicName + " does not exist");
-		}
-
-		try {
-			binder.getConnectionFactory().getPartitions(testTopicName);
-			fail();
-		}
-		catch (Exception e) {
-			assertThat(e).isInstanceOf(TopicNotFoundException.class);
-		}
-	}
-
-	@Test
-	public void testAutoConfigureTopicsDisabledSucceedsIfTopicExisting() throws Exception {
-		String testTopicName = "existing" + System.currentTimeMillis();
-		AdminUtils.createTopic(kafkaTestSupport.getZkClient(), testTopicName, 5, 1, new Properties());
-		KafkaBinderConfigurationProperties configurationProperties = createConfigurationProperties();
-		configurationProperties.setAutoCreateTopics(false);
-		KafkaMessageChannelBinder binder = new KafkaMessageChannelBinder(configurationProperties);
-		GenericApplicationContext context = new GenericApplicationContext();
-		context.refresh();
-		binder.setApplicationContext(context);
-		binder.afterPropertiesSet();
-		ExtendedConsumerProperties<KafkaConsumerProperties> consumerProperties = createConsumerProperties();
-		DirectChannel output = createBindableChannel("output", createConsumerBindingProperties(consumerProperties));
-		Binding<MessageChannel> binding = binder.doBindConsumer(testTopicName, "test", output, consumerProperties);
-		binding.unbind();
-	}
-
-	@Test
-	public void testAutoAddPartitionsDisabledFailsIfTopicUnderpartitioned() throws Exception {
-		String testTopicName = "existing" + System.currentTimeMillis();
-		AdminUtils.createTopic(kafkaTestSupport.getZkClient(), testTopicName, 1, 1, new Properties());
-		KafkaBinderConfigurationProperties configurationProperties = createConfigurationProperties();
-		configurationProperties.setAutoAddPartitions(false);
-		KafkaMessageChannelBinder binder = new KafkaMessageChannelBinder(configurationProperties);
-		GenericApplicationContext context = new GenericApplicationContext();
-		context.refresh();
-		binder.setApplicationContext(context);
-		binder.afterPropertiesSet();
-		ExtendedConsumerProperties<KafkaConsumerProperties> consumerProperties = createConsumerProperties();
-		// this consumer must consume from partition 2
-		consumerProperties.setInstanceCount(3);
-		consumerProperties.setInstanceIndex(2);
-		DirectChannel output = createBindableChannel("output", createConsumerBindingProperties(consumerProperties));
-		try {
-			binder.doBindConsumer(testTopicName, "test", output, consumerProperties);
-		}
-		catch (Exception e) {
-			assertThat(e).isInstanceOf(BinderException.class);
-			assertThat(e).hasMessageContaining(
-					"The number of expected partitions was: 3, but 1 has been found instead");
-		}
-	}
-
-	@Test
-	public void testAutoAddPartitionsDisabledSucceedsIfTopicPartitionedCorrectly() throws Exception {
-
-		String testTopicName = "existing" + System.currentTimeMillis();
-		AdminUtils.createTopic(kafkaTestSupport.getZkClient(), testTopicName, 6, 1, new Properties());
-		KafkaBinderConfigurationProperties configurationProperties = createConfigurationProperties();
-		configurationProperties.setAutoAddPartitions(false);
-		KafkaMessageChannelBinder binder = new KafkaMessageChannelBinder(configurationProperties);
-		GenericApplicationContext context = new GenericApplicationContext();
-		RetryTemplate metatadataRetrievalRetryOperations = new RetryTemplate();
-		metatadataRetrievalRetryOperations.setRetryPolicy(new SimpleRetryPolicy());
-		FixedBackOffPolicy backOffPolicy = new FixedBackOffPolicy();
-		backOffPolicy.setBackOffPeriod(1000);
-		metatadataRetrievalRetryOperations.setBackOffPolicy(backOffPolicy);
-		binder.setMetadataRetryOperations(metatadataRetrievalRetryOperations);
-		context.refresh();
-		binder.setApplicationContext(context);
-		binder.afterPropertiesSet();
-		ExtendedConsumerProperties<KafkaConsumerProperties> consumerProperties = createConsumerProperties();
-		// this consumer must consume from partition 2
-		consumerProperties.setInstanceCount(3);
-		consumerProperties.setInstanceIndex(2);
-		DirectChannel output = createBindableChannel("output", createConsumerBindingProperties(consumerProperties));
-		Binding<?> binding = binder.doBindConsumer(testTopicName, "test", output, consumerProperties);
-
-		Partition[] listenedPartitions = TestUtils.getPropertyValue(binding,
-				"endpoint.messageListenerContainer.partitions", Partition[].class);
-
-		assertThat(listenedPartitions).hasSize(2);
-		assertThat(listenedPartitions).contains(new Partition(testTopicName, 2), new Partition(testTopicName, 5));
-		Collection<Partition> partitions = binder.getConnectionFactory().getPartitions(testTopicName);
-		assertThat(partitions).hasSize(6);
-		binding.unbind();
-	}
-
-	@Test
-	public void testAutoCreateTopicsEnabledSucceeds() throws Exception {
-		KafkaBinderConfigurationProperties configurationProperties = createConfigurationProperties();
-		configurationProperties.setAutoCreateTopics(true);
-		KafkaMessageChannelBinder binder = new KafkaMessageChannelBinder(configurationProperties);
-		GenericApplicationContext context = new GenericApplicationContext();
-		context.refresh();
-		binder.setApplicationContext(context);
-		binder.afterPropertiesSet();
-		RetryTemplate metatadataRetrievalRetryOperations = new RetryTemplate();
-		metatadataRetrievalRetryOperations.setRetryPolicy(new SimpleRetryPolicy());
-		FixedBackOffPolicy backOffPolicy = new FixedBackOffPolicy();
-		backOffPolicy.setBackOffPeriod(1000);
-		metatadataRetrievalRetryOperations.setBackOffPolicy(backOffPolicy);
-		binder.setMetadataRetryOperations(metatadataRetrievalRetryOperations);
-		ExtendedConsumerProperties<KafkaConsumerProperties> consumerProperties = createConsumerProperties();
-		DirectChannel output = createBindableChannel("output", createConsumerBindingProperties(consumerProperties));
-		String testTopicName = "nonexisting" + System.currentTimeMillis();
-		Binding<?> binding = binder.doBindConsumer(testTopicName, "test", output, consumerProperties);
-		binding.unbind();
-	}
-
-	@Test
-	public void testPartitionCountNotReduced() throws Exception {
-		String testTopicName = "existing" + System.currentTimeMillis();
-		AdminUtils.createTopic(kafkaTestSupport.getZkClient(), testTopicName, 6, 1, new Properties());
-		KafkaBinderConfigurationProperties configurationProperties = createConfigurationProperties();
-		configurationProperties.setAutoAddPartitions(true);
-		KafkaMessageChannelBinder binder = new KafkaMessageChannelBinder(configurationProperties);
-		GenericApplicationContext context = new GenericApplicationContext();
-		context.refresh();
-		binder.setApplicationContext(context);
-		binder.afterPropertiesSet();
-		RetryTemplate metatadataRetrievalRetryOperations = new RetryTemplate();
-		metatadataRetrievalRetryOperations.setRetryPolicy(new SimpleRetryPolicy());
-		FixedBackOffPolicy backOffPolicy = new FixedBackOffPolicy();
-		backOffPolicy.setBackOffPeriod(1000);
-		metatadataRetrievalRetryOperations.setBackOffPolicy(backOffPolicy);
-		binder.setMetadataRetryOperations(metatadataRetrievalRetryOperations);
-		DirectChannel output = new DirectChannel();
-		ExtendedConsumerProperties<KafkaConsumerProperties> consumerProperties = createConsumerProperties();
-		Binding<?> binding = binder.doBindConsumer(testTopicName, "test", output, consumerProperties);
-		binding.unbind();
-		TopicMetadata topicMetadata = AdminUtils.fetchTopicMetadataFromZk(testTopicName,
-				kafkaTestSupport.getZkClient());
-		assertThat(topicMetadata.partitionsMetadata().size()).isEqualTo(6);
-	}
-
-	@Test
-	public void testPartitionCountIncreasedIfAutoAddPartitionsSet() throws Exception {
-		String testTopicName = "existing" + System.currentTimeMillis();
-		AdminUtils.createTopic(kafkaTestSupport.getZkClient(), testTopicName, 1, 1, new Properties());
-		KafkaBinderConfigurationProperties configurationProperties = createConfigurationProperties();
-		configurationProperties.setMinPartitionCount(6);
-		configurationProperties.setAutoAddPartitions(true);
-		KafkaMessageChannelBinder binder = new KafkaMessageChannelBinder(configurationProperties);
-		GenericApplicationContext context = new GenericApplicationContext();
-		context.refresh();
-		binder.setApplicationContext(context);
-		binder.afterPropertiesSet();
-		RetryTemplate metatadataRetrievalRetryOperations = new RetryTemplate();
-		metatadataRetrievalRetryOperations.setRetryPolicy(new SimpleRetryPolicy());
-		FixedBackOffPolicy backOffPolicy = new FixedBackOffPolicy();
-		backOffPolicy.setBackOffPeriod(1000);
-		metatadataRetrievalRetryOperations.setBackOffPolicy(backOffPolicy);
-		binder.setMetadataRetryOperations(metatadataRetrievalRetryOperations);
-		DirectChannel output = new DirectChannel();
-		ExtendedConsumerProperties<KafkaConsumerProperties> consumerProperties = createConsumerProperties();
-		Binding<?> binding = binder.doBindConsumer(testTopicName, "test", output, consumerProperties);
-		binding.unbind();
-		TopicMetadata topicMetadata = AdminUtils.fetchTopicMetadataFromZk(testTopicName,
-				kafkaTestSupport.getZkClient());
-		assertThat(topicMetadata.partitionsMetadata().size()).isEqualTo(6);
-	}
-
-	KafkaProducerProperties.CompressionType fromProducerMetadataCompressionType(
-			ProducerMetadata.CompressionType compressionType) {
-		switch  (compressionType) {
-			case snappy : return KafkaProducerProperties.CompressionType.snappy;
-			case gzip	: return KafkaProducerProperties.CompressionType.gzip;
-			default: return KafkaProducerProperties.CompressionType.none;
-		}
-	}
-
-	private static final class FailingInvocationCountingMessageHandler implements MessageHandler {
-
-		private int invocationCount;
-
-		private final LinkedHashMap<Long, Message<?>> receivedMessages = new LinkedHashMap<>();
-
-		private final CountDownLatch latch;
-
-		private FailingInvocationCountingMessageHandler(int latchSize) {
-			latch = new CountDownLatch(latchSize);
-		}
-
-		private FailingInvocationCountingMessageHandler() {
-			this(1);
-		}
-
-		@Override
-		public void handleMessage(Message<?> message) throws MessagingException {
-			invocationCount++;
-			Long offset = message.getHeaders().get(KafkaHeaders.OFFSET, Long.class);
-			// using the offset as key allows to ensure that we don't store duplicate
-			// messages on retry
-			if (!receivedMessages.containsKey(offset)) {
-				receivedMessages.put(offset, message);
-				latch.countDown();
-			}
-			throw new RuntimeException();
-		}
-
-		public LinkedHashMap<Long, Message<?>> getReceivedMessages() {
-			return receivedMessages;
-		}
-
-		public int getInvocationCount() {
-			return invocationCount;
-		}
-
-		public CountDownLatch getLatch() {
-			return latch;
-		}
-	}
-}
